+++ src/lib/components/ExplorerLink.tsx
@@ -324,7 +324,7 @@ export const ExplorerLink = ({
       borderWidth="1px"
       display="inline-flex"
       gap={1}
-      h={fixedHeight ? "24px" : "auto"}
+      h={fixedHeight && textFormat !== "normal" ? "24px" : "auto"}
       maxW={textLabel ? "100%" : "fit-content"}
       px={1}
       rounded={4}
@@ -343,14 +343,12 @@ export const ExplorerLink = ({
       w="fit-content"
       onMouseEnter={() => setHoveredText(value)}
       onMouseLeave={() => setHoveredText(null)}
+      onTouchEnd={() => setHoveredText(null)}
+      onTouchStart={() => setHoveredText(value)}
       {...componentProps}
     >
       {leftIcon}
-      <Tooltip
-        hidden={isTooltipHidden(type, textFormat)}
-        label={value}
-        textAlign="center"
-      >
+      {isMobile ? (
         <LinkRender
           chainId={chainId}
           fallbackValue={copyValue ?? ""}
@@ -363,7 +361,26 @@ export const ExplorerLink = ({
           textVariant={textVariant}
           type={type}
         />
-      </Tooltip>
+      ) : (
+        <Tooltip
+          hidden={isTooltipHidden(type, textFormat)}
+          label={value}
+          textAlign="center"
+        >
+          <LinkRender
+            chainId={chainId}
+            fallbackValue={copyValue ?? ""}
+            hrefLink={link}
+            isEllipsis={textFormat === "ellipsis"}
+            isInternal={isUndefined(externalLink)}
+            openNewTab={openNewTab}
+            textFormat={textFormat}
+            textValue={textValue}
+            textVariant={textVariant}
+            type={type}
+          />
+        </Tooltip>
+      )}
       {rightIcon}
       {!hideCopy && (
         <Copier