+++ CHANGELOG.md
@@ -50,6 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#1443](https://github.com/alleslabs/celatone-frontend/pull/1443) Update ExplorerLink component for mobile users
 - [#1405](https://github.com/alleslabs/celatone-frontend/pull/1405) Enhance Readme.md and add CODEOWNERS
 
 ### Bug fixes